Weapons: Added: Upgrade damage and Power, both scale off of weapon prestige level and now reset at start of the game

+++ Assets/Scripts/Enemy.cs
@@ -5,8 +5,8 @@ using UnityEngine;
 
 public class Enemy : MonoBehaviour
 {
-    public int maxHealth = 100;
-    public int currentHealth;
+    public float maxHealth = 100;
+    public float currentHealth;
 
     public float Damage = 10.0f;
     public float MovementSpeed = 5f;
@@ -60,7 +60,7 @@ public class Enemy : MonoBehaviour
         }
     }
 
-    public void TakeDamage(int damage)
+    public void TakeDamage(float damage)
     {
         currentHealth -= damage;
         healthBar.SetHealth(currentHealth);
